+++ Demo/Tests/MoyaProviderSpec.swift
@@ -348,6 +348,65 @@ class MoyaProviderSpec: QuickSpec {
                 expect(executed).to(beTruthy())
             }
         }
+
+        describe("a provider with custom sample response closures") {
+            it("returns sample data") {
+                let endpointResolution: MoyaProvider<GitHub>.EndpointClosure = { target in
+                    let url = target.baseURL.appendingPathComponent(target.path).absoluteString
+                    return Endpoint(URL: url, sampleResponseClosure: {.networkResponse(200, target.sampleData)}, method: target.method, parameters: target.parameters)
+                }
+                let provider = MoyaProvider<GitHub>(endpointClosure: endpointResolution, stubClosure: MoyaProvider.ImmediatelyStub)
+
+                var data: Data?
+                _ = provider.request(GitHub.zen) { result in
+                    if case .success(let response) = result {
+                        data = response.data
+                    }
+                }
+
+                expect(data) == GitHub.zen.sampleData
+            }
+
+            it("returns identical sample response") {
+                let response = HTTPURLResponse(url: URL(string: "http://example.com")!, mimeType: nil, expectedContentLength: 0, textEncodingName: nil)
+                let endpointResolution: MoyaProvider<GitHub>.EndpointClosure = { target in
+                    let url = target.baseURL.appendingPathComponent(target.path).absoluteString
+                    return Endpoint(URL: url, sampleResponseClosure: { .response(response, Data()) }, method: target.method, parameters: target.parameters)
+                }
+                let provider = MoyaProvider<GitHub>(endpointClosure: endpointResolution, stubClosure: MoyaProvider.ImmediatelyStub)
+
+                var receivedResponse: URLResponse?
+                _ = provider.request(GitHub.zen) { result in
+                    if case .success(let response) = result {
+                        receivedResponse = response.response
+                    }
+                }
+
+                expect(receivedResponse) === response
+            }
+
+            it("returns error") {
+                let error = NSError(domain: "Internal iOS Error", code: -1234, userInfo: nil)
+                let endpointResolution: MoyaProvider<GitHub>.EndpointClosure = { target in
+                    let url = target.baseURL.appendingPathComponent(target.path).absoluteString
+                    return Endpoint(URL: url, sampleResponseClosure: { .networkError(error) }, method: target.method, parameters: target.parameters)
+                }
+                let provider = MoyaProvider<GitHub>(endpointClosure: endpointResolution, stubClosure: MoyaProvider.ImmediatelyStub)
+
+                var receivedError: Moya.Error?
+                _ = provider.request(GitHub.zen) { result in
+                    if case .failure(let error) = result {
+                        receivedError = error
+                    }
+                }
+
+                if case .some(Moya.Error.underlying(let underlyingError as NSError)) = receivedError {
+                    expect(underlyingError) == error
+                } else {
+                    fail("Expected to receive error, did not.")
+                }
+            }
+        }
         
         describe("a provider with error in request closure") {
             var provider: MoyaProvider<GitHub>!
@@ -585,6 +644,52 @@ class MoyaProviderSpec: QuickSpec {
                 }
                 
                 expect(underlyingIsCancelled).to(beTrue())
+            }
+        }
+        
+        describe("a provider with progress tracking") {
+            var provider: MoyaProvider<GitHubUserContent>!
+            beforeEach {
+                
+                //delete downloaded filed before each test
+                let directoryURLs = FileManager.default.urls(for: .cachesDirectory, in: .userDomainMask)
+                let file = directoryURLs.first!.appendingPathComponent("logo_github.png")
+                try? FileManager.default.removeItem(at: file)
+                
+                //`responseTime(-4)` equals to 1000 bytes at a time. The sample data is 4000 bytes.
+                OHHTTPStubs.stubRequests(passingTest: {$0.url!.path.hasSuffix("logo_github.png")}) { _ in
+                    return OHHTTPStubsResponse(data: GitHubUserContent.downloadMoyaWebContent("logo_github.png").sampleData, statusCode: 200, headers: nil).responseTime(-4)
+                }
+                provider = MoyaProvider<GitHubUserContent>()
+            }
+            
+            it("tracks progress of request") {
+                
+                let target: GitHubUserContent = .downloadMoyaWebContent("logo_github.png")
+                
+                var progressValues: [Double] = []
+                var completedValues: [Bool] = []
+                var error: Moya.Error?
+                
+                waitUntil(timeout: 5.0) { done in
+                    let progressClosure: ProgressBlock = { progress in
+                        progressValues.append(progress.progress)
+                        completedValues.append(progress.completed)
+                    }
+                    
+                    let progressCompletionClosure: Completion = { (result) in
+                        if case .failure(let err) = result {
+                            error = err
+                        }
+                        done()
+                    }
+                    
+                    provider.request(target, queue: nil, progress: progressClosure, completion: progressCompletionClosure)
+                }
+                
+                expect(error).to(beNil())
+                expect(progressValues) == [0.25, 0.5, 0.75, 1.0, 1.0]
+                expect(completedValues) == [false, false, false, false, true]
             }
         }
     }