+++ Source/Moya+Alamofire.swift
@@ -61,3 +61,25 @@ public final class CancellableToken: Cancellable, CustomDebugStringConvertible {
     }
 
 }
+
+internal typealias RequestableCompletion = (HTTPURLResponse?, URLRequest?, Data?, Swift.Error?) -> Void
+
+internal protocol Requestable {
+    func response(queue: DispatchQueue?, completionHandler: @escaping RequestableCompletion) -> Self
+}
+
+extension DataRequest: Requestable {
+    internal func response(queue: DispatchQueue?, completionHandler: @escaping RequestableCompletion) -> Self {
+        return response(queue: queue, completionHandler: { handler  in
+            completionHandler(handler.response, handler.request, handler.data, handler.error)
+        })
+    }
+}
+
+extension DownloadRequest: Requestable {
+    internal func response(queue: DispatchQueue?, completionHandler: @escaping RequestableCompletion) -> Self {
+        return response(queue: queue, completionHandler: { handler  in
+            completionHandler(handler.response, handler.request, nil, handler.error)
+        })
+    }
+}