+++ Changelog.md
@@ -9,6 +9,10 @@
 - **Breaking Change** Changed HTTP verbs enum to lowercase
 - `urlRequest` property of `Endpoint` is now truly optional. The request will fail if the `urlRequest` turns out to be nil and a `requestMapping` error will be returned together with the problematic url.
 - **Breaking Change** Made RxMoya & ReactiveMoya frameworks dependant on Moya framework, making them slimmer and not re-including Moya source in the Reactive extensions. ([PR](https://github.com/Moya/Moya/pull/563))
+- Fixes download requests never calling the completion block.
+- Adds new internal Requestable protocol.
+- Added a new case to `SampleResponseClosure` which allows mocking of the whole `URLResponse`.
+- Adds test for new `SampleResponseClosure` case.
 
 # 8.0.0-beta.2